requirements.txt
pymongo==3.9.0
celery==4.3.0
redis==3.3.11
python-nmap==0.6.1
tldextract==2.2.2
dnspython==1.16.0
python-libnmap==0.7.0
pika==1.1.0
nmap==0.0.1
tldextract==2.2.2
dnspython==1.16.0
beautifulsoup4==4.8.2
geoip2==3.0.0
flask-restful==0.3.8
gunicorn==20.0.4
flask_cors==3.0.8
requests==2.23.0